Changed how location is made
So, before, there was nothing, and on the first day, I used the google
geocode nuget package to receive longitude and latitude and i saw that
it was good

on the second day, i realized i had made a grievous error in that I was
not sending back a human readable version of the address i looked up.

ie, if you were to pass in Pleasanton or something into the address, how
many cities are there called pleasanton (hopefully more than one... for
this example...) so i wouldn't know what longitude and latitude was
being sent back.

needless to say, i had to recreate some stuff and now the code looks
awful and i'll update it to be cleaner next time.

integration tests passed though
/shrug