quick refactor to not return tuples
the tuple return with an error message was a relic of when the weather
wervice also used the location service inside of itself to get the
location information. now that's happening at the controller level so
now i can make error messages based on just which service failed inside
of the controller. no need for the service to handle that logic and
return a tuple